[ML] move declarations
